Factory function created for app set up with namespaces added. Run file to run the app. Routes transferred to namespaces
